+++ Source/default.py
@@ -2,13 +2,20 @@ import xbmc
 import xbmcgui
 import xbmcaddon
 import metaData
-import install
+import jsongets
+import searchOptions
+import searchProfile
+# import kesGUI
 
-install.install()
+searchProfile.install()
 
 addon = xbmcaddon.Addon()
 
 ##
+
+##
+
+# interface = kesGUI.build()
 
 
 class MyPlayer(xbmcgui.Window):
@@ -35,5 +42,28 @@ if (__name__ == "__main__"):
     w.doModal()
     del w
 """
-metaData.testMetaData()
+
+# **********************
+'''
+This is a function where we can compile all of our unit tests
+Each Python file should have a single unit test function to test all other functions
+'''
+# **********************
+
+
+def unitTest():
+    # Meta Data Unit Tests
+    # Prints the results of each of the get meta data functions
+    metaData.testMetaData()
+
+    # Search Option Unit Test
+    # Creates a searchOption object, sets, adds, and removes each potential option and prints results, will finish by returning the options object
+    options = searchOptions.unitTest()
+
+    # Search Profile Unit Test
+    searchProfile.unitTest(options)
+
+
+unitTest()
+
 xbmcgui.Dialog().ok("KES", "Complete!")