+++ Source/searchOptions.py
@@ -289,5 +289,4 @@ def unitTest():
     testCase.setPBFunction(2)
     print(testCase.getPBFunction())
 
-
-unitTest()
+    return testCase